+++ lib/src/ad_types.dart
@@ -7,9 +7,12 @@ class AdTypeFlags {
   static const int appOpen = 1 << 6;
 
   @Deprecated("Use AdTypeFlags.banner instead")
+  // ignore: constant_identifier_names
   static const int Banner = banner;
   @Deprecated("Use AdTypeFlags.interstitial instead")
+  // ignore: constant_identifier_names
   static const int Interstitial = interstitial;
   @Deprecated("Use AdTypeFlags.rewarded instead")
+  // ignore: constant_identifier_names
   static const int Rewarded = rewarded;
 }