+++ lib/src/audience.dart
@@ -3,9 +3,12 @@ enum Audience {
   children,
   notChildren,
   @Deprecated("Use Audience.undefined instead")
+  // ignore: constant_identifier_names
   UNDEFINED,
   @Deprecated("Use Audience.children instead")
+  // ignore: constant_identifier_names
   CHILDREN,
   @Deprecated("Use Audience.notChildren instead")
+  // ignore: constant_identifier_names
   NOT_CHILDREN
 }