+++ lib/src/banner/ad_position.dart
@@ -9,15 +9,21 @@ enum AdPosition {
   middleLeft,
   middleRight,
   @Deprecated("Use AdPosition.topCenter instead")
+  // ignore: constant_identifier_names
   TopCenter,
   @Deprecated("Use AdPosition.topLeft instead")
+  // ignore: constant_identifier_names
   TopLeft,
   @Deprecated("Use AdPosition.topRight instead")
+  // ignore: constant_identifier_names
   TopRight,
   @Deprecated("Use AdPosition.bottomCenter instead")
+  // ignore: constant_identifier_names
   BottomCenter,
   @Deprecated("Use AdPosition.bottomLeft instead")
+  // ignore: constant_identifier_names
   BottomLeft,
   @Deprecated("Use AdPosition.bottomRight instead")
+  // ignore: constant_identifier_names
   BottomRight
 }