+++ lib/src/banner/internal/internal_listener_container.dart
@@ -19,30 +19,36 @@ class InternalListenerContainer {
   Future<dynamic> handleMethodCall(MethodCall call) async {
     switch (call.method) {
       case 'OnBannerAdLoaded':
+        // ignore: deprecated_member_use_from_same_package
         getBannerListener(call)?.onLoaded();
         break;
 
       case 'OnBannerAdShown':
+        // ignore: deprecated_member_use_from_same_package
         getBannerListener(call)?.onAdViewPresented();
         break;
 
       case 'OnBannerAdImpression':
+        // ignore: deprecated_member_use_from_same_package
         getBannerListener(call)?.onImpression(AdImpression.tryParse(call));
         break;
 
       case 'OnBannerAdFailedToShow':
         if ((call.arguments as Object?) != null) {
+          // ignore: deprecated_member_use_from_same_package
           getBannerListener(call)?.onFailed(call.arguments["message"]);
         }
         break;
 
       case 'OnBannerAdFailedToLoad':
         if ((call.arguments as Object?) != null) {
+          // ignore: deprecated_member_use_from_same_package
           getBannerListener(call)?.onFailed(call.arguments["message"]);
         }
         break;
 
       case 'OnBannerAdClicked':
+        // ignore: deprecated_member_use_from_same_package
         getBannerListener(call)?.onClicked();
         break;
     }