+++ lib/src/cas.dart
@@ -17,7 +17,7 @@ import 'user_consent.dart';
 
 /// Represents the CAS.AI SDK.
 class CAS {
-  static const String _pluginVersion = "0.8.2";
+  static const String _pluginVersion = "0.8.3";
 
   static const MethodChannel _channel = MethodChannel("cleveradssolutions/cas");
 
@@ -51,7 +51,8 @@ class CAS {
   }
 
   static Future<String> getSDKVersion() async {
-    final String? sdkVersion = await _channel.invokeMethod<String>('getSDKVersion');
+    final String? sdkVersion =
+        await _channel.invokeMethod<String>('getSDKVersion');
     return sdkVersion ?? '';
   }