+++ lib/src/ccpa_status.dart
@@ -3,9 +3,12 @@ enum CCPAStatus {
   optOutSale,
   optInSale,
   @Deprecated("Use CCPAStatus.undefined instead")
+  // ignore: constant_identifier_names
   UNDEFINED,
   @Deprecated("Use CCPAStatus.optOutSale instead")
+  // ignore: constant_identifier_names
   OPT_OUT_SALE,
   @Deprecated("Use CCPAStatus.optInSale instead")
+  // ignore: constant_identifier_names
   OPT_IN_SALE
 }