+++ lib/src/loading_mode.dart
@@ -1,8 +1,14 @@
 enum LoadingMode {
+  // ignore: constant_identifier_names
   FastestRequests,
+  // ignore: constant_identifier_names
   FastRequests,
+  // ignore: constant_identifier_names
   Optimal,
+  // ignore: constant_identifier_names
   HighPerformance,
+  // ignore: constant_identifier_names
   HighestPerformance,
+  // ignore: constant_identifier_names
   Manual
 }