+++ lib/src/user_consent.dart
@@ -1,2 +1,3 @@
 @Deprecated("Use ConsentStatus")
+// ignore: constant_identifier_names
 enum UserConsent { UNDEFINED, ACCEPTED, DENIED }